# Feike/malloc performance improvements