--- a/crates/udd-sketch/src/compactor.rs
+++ b/crates/udd-sketch/src/compactor.rs
@@ -1,30 +1,32 @@
+use crate::swap::SwapBucket;
 use crate::SketchHashKey::Invalid;
 use crate::{SketchHashEntry, SketchHashKey, SketchHashMap};
 
-type T = (SketchHashKey, u64);
-
 #[inline]
-pub fn compact_from_iter(swap_iter: &mut impl Iterator<Item = T>, map: &mut SketchHashMap) {
+pub fn compact_from_iter(
+    swap_iter: &mut impl Iterator<Item = SwapBucket>,
+    map: &mut SketchHashMap,
+) {
     let Some(mut current) = swap_iter.next() else {
         return;
     };
 
     for next in swap_iter {
-        if next.0 == Invalid {
+        if next.key == Invalid {
             break;
         }
 
         // This combines those buckets that compact into the same one
         // For example, Positive(9) and Positive(8) both
         // compact into Positive(4)
-        if current.0 == next.0 {
-            current.1 += next.1;
+        if current.key == next.key {
+            current.count += next.count;
         } else {
             map.map.insert(
-                current.0,
+                current.key,
                 SketchHashEntry {
-                    count: current.1,
-                    next: next.0,
+                    count: current.count,
+                    next: next.key,
                 },
             );
             current = next;
@@ -33,9 +35,9 @@ pub fn compact_from_iter(swap_iter: &mut impl Iterator<Item = T>, map: &mut Sket
 
     // And the final one ...
     map.map.insert(
-        current.0,
+        current.key,
         SketchHashEntry {
-            count: current.1,
+            count: current.count,
             next: Invalid,
         },
     );