--- a/extension/Cargo.toml
+++ b/extension/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "timescaledb_toolkit"
-version = "1.19.0-dev"
+version = "1.20.0-dev"
 edition = "2021"
 
 [[bin]]