@feikesteenbergen 2025-04-10
Closing in favor of the simpler https://github.com/timescale/timescaledb-toolkit/pull/853